refactor(podman): grouping check(s) files
