+++ extensions/podman/packages/extension/src/checks/detection-checks.ts
@@ -18,8 +18,8 @@
 
 import type * as extensionApi from '@podman-desktop/api';
 
-import type { InstalledPodman } from './podman-cli';
-import { getCustomBinaryPath, getInstallationPath } from './podman-cli';
+import type { InstalledPodman } from '../podman-cli';
+import { getCustomBinaryPath, getInstallationPath } from '../podman-cli';
 
 export function getDetectionChecks(installedPodman?: InstalledPodman): extensionApi.ProviderDetectionCheck[] {
   const detectionChecks: extensionApi.ProviderDetectionCheck[] = [];
