+++ extensions/podman/packages/extension/src/checks/macos-check.spec.ts
