+++ extensions/podman/packages/extension/src/checks/macos-checks.ts
